search_links_scraper/requirements.txt
aiofiles==23.2.1
anyio==3.7.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
cachetools==5.3.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
customtkinter==5.2.0
darkdetect==0.8.0
duckduckgo-search==3.9.5
duckpy==3.2.0
filelock==3.13.1
gdown==4.7.1
google==3.0.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-search-results==2.4.2
googleapis-common-protos==1.61.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.13.7
httplib2==0.22.0
httpx==0.18.2
hyperframe==5.2.0
idna==3.4
joblib==1.3.2
lxml==4.9.3
nltk==3.8.1
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDrive==1.3.1
pyparsing==3.0.9
PySocks==1.7.1
python-decouple==3.8
python-dotenv==0.21.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.11.3
selenium==4.6.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
